refactor microservices

+++ ms-customers/src/main/java/com/sofka/ms_customers/service/CustomerServiceImpl.java
@@ -3,12 +3,14 @@ package com.sofka.ms_customers.service;
 import com.sofka.ms_customers.dto.CustomerDTO;
 import com.sofka.ms_customers.entity.Customer;
 import com.sofka.ms_customers.entity.Person;
+import com.sofka.ms_customers.exceptions.CustomerNotFoundException;
 import com.sofka.ms_customers.mapper.CustomerMapper;
 import com.sofka.ms_customers.repository.CustomerRepository;
 import com.sofka.ms_customers.repository.PersonRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,13 +23,7 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public CustomerDTO createCustomer(CustomerDTO customerDTO) {
-        Person person = new Person();
-        person.setName(customerDTO.getName());
-        person.setGender(customerDTO.getGender());
-        person.setAge(customerDTO.getAge());
-        person.setIdentification(customerDTO.getIdentification());
-        person.setAddress(customerDTO.getAddress());
-        person.setPhone(customerDTO.getPhone());
+        Person person = createPersonFromDTO(customerDTO);
         person = personRepository.save(person);
 
         Customer customer = customerMapper.toEntityWithPerson(customerDTO);
@@ -38,11 +34,34 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
+    public CustomerDTO updateCustomer(CustomerDTO customerDTO) {
+        Customer existingCustomer = customerRepository.findById(customerDTO.getCustomerId())
+                .orElseThrow(() -> new CustomerNotFoundException("Cliente no encontrado con identificación: " + customerDTO.getCustomerId()));
+
+        Person person = existingCustomer.getPerson();
+
+        if (customerDTO.getName() != null) person.setName(customerDTO.getName());
+        if (customerDTO.getGender() != null) person.setGender(customerDTO.getGender());
+        if (customerDTO.getAge() != 0) person.setAge(customerDTO.getAge());
+        if (customerDTO.getIdentification() != null) person.setIdentification(customerDTO.getIdentification());
+        if (customerDTO.getAddress() != null) person.setAddress(customerDTO.getAddress());
+        if (customerDTO.getPhone() != null) person.setPhone(customerDTO.getPhone());
+
+        personRepository.save(person);
+
+        if (customerDTO.getPassword() != null) existingCustomer.setPassword(customerDTO.getPassword());
+        if (customerDTO.getStatus() != null) existingCustomer.setStatus(customerDTO.getStatus());
+
+        Customer updatedCustomer = customerRepository.save(existingCustomer);
+
+        return customerMapper.toDTO(updatedCustomer);
+    }
+
+    @Override
     public CustomerDTO getCustomerByIdentification(String identification) {
-        System.out.println("LLega: "+identification);
         return customerRepository.findByIdentification(identification)
                 .map(customerMapper::toDTO)
-                .orElseThrow(() -> new RuntimeException("Cliente no encontrado"));
+                .orElseThrow(() -> new CustomerNotFoundException("Cliente no encontrado con identificación: " + identification));
     }
 
     @Override
@@ -53,7 +72,21 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public void deleteCustomer(String customerId) {
+    public void deleteCustomer(UUID customerId) {
+        if (!customerRepository.existsById(customerId)) {
+            throw new CustomerNotFoundException("Cliente no encontrado con ID: " + customerId);
+        }
         customerRepository.deleteById(customerId);
+    }
+
+    private Person createPersonFromDTO(CustomerDTO customerDTO) {
+        Person person = new Person();
+        person.setName(customerDTO.getName());
+        person.setGender(customerDTO.getGender());
+        person.setAge(customerDTO.getAge());
+        person.setIdentification(customerDTO.getIdentification());
+        person.setAddress(customerDTO.getAddress());
+        person.setPhone(customerDTO.getPhone());
+        return person;
     }
 }
